+++ uhppoted/devices.go
@@ -104,10 +104,10 @@ func (u *UHPPOTED) GetDevice(request GetDeviceRequest) (*GetDeviceResponse, erro
 	return &response, nil
 }
 
-func (u *UHPPOTED) SetEventListener(controller uint32, addr netip.AddrPort) (bool, error) {
-	u.debug("set-event-listener", fmt.Sprintf("%v %v", controller, addr))
+func (u *UHPPOTED) SetEventListener(controller uint32, addr netip.AddrPort, interval uint8) (bool, error) {
+	u.debug("set-event-listener", fmt.Sprintf("%v address:%v interval:%v", controller, addr, interval))
 
-	if ok, err := u.UHPPOTE.SetListener(controller, addr, 0); err != nil {
+	if ok, err := u.UHPPOTE.SetListener(controller, addr, interval); err != nil {
 		return false, fmt.Errorf("%w: %v", ErrInternalServerError, fmt.Errorf("set-event-listener: %v %w", controller, err))
 	} else if !ok {
 		return false, fmt.Errorf("%w: %v", ErrInternalServerError, fmt.Errorf("set-event-listener: %v  failed", controller))